Synx Project, remove next groups of Switflint warnings.
