+++ Sources/Framework/Business Logic/Alerts/AlertActionConfiguration.swift
@@ -1,0 +1,19 @@
+import UIKit
+
+public struct AlertActionConfiguration {
+
+    let title: String
+    let style: UIAlertActionStyle
+    let handler: (() -> Void)?
+
+}
+
+// MARK: - Equatables
+
+extension AlertActionConfiguration: Equatable {
+
+    public static func == (lhs: AlertActionConfiguration, rhs: AlertActionConfiguration) -> Bool {
+        return lhs.title == rhs.title && lhs.style == rhs.style
+    }
+
+}
