+++ Sources/Framework/Extensions/UIView+Sizes.swift
@@ -2,7 +2,7 @@ import UIKit
 
 extension UIView {
 
-    public var height: CGFloat {
+     var height: CGFloat {
         return self.bounds.size.height
     }
