+++ Sources/Framework/Models/Debate.swift
@@ -14,7 +14,7 @@ extension Debate {
 
     public var lastAnswerType: AnswerType? {
         let answers: [Answer] = [positiveAnswer, neutralAnswer, negativeAnswer]
-        return answers.first(where: { $0.identifier == lastAnswerIdentifier })?.type
+        return answers.first { $0.identifier == lastAnswerIdentifier }?.type
     }
 
 }
