+++ Sources/Framework/Networking/ApiClient.swift
@@ -52,7 +52,7 @@ class ApiClient: APIProviding {
             headers: ["Authorization": authToken]
         )
 
-        return Promise(requestExecutor: response.maybeJson, processor: { _ in answer })
+        return Promise(requestExecutor: response.maybeJson) { _ in answer }
             .recover { error -> Promise<Answer> in
                 if case let AFError.responseValidationFailed(reason: .unacceptableStatusCode(code: code)) = error,
                     code == 429 {
@@ -71,14 +71,14 @@ class ApiClient: APIProviding {
             headers: ["Authorization": voteContext.authToken]
         )
 
-        return Promise(requestExecutor: response.maybeJson, processor: { _ in true })
+        return Promise(requestExecutor: response.maybeJson) { _ in true }
     }
 
 }
 
 private func request<T>(with executor: @escaping (@escaping (ApiResponse) -> Void) -> Void,
                         deserializedBy deserializer: Deserializer<T>) -> Promise<T> {
-    return Promise(requestExecutor: executor, processor: { response in
+    return Promise(requestExecutor: executor) { response in
         try deserializer.deserialize(json: response.json)
-    })
+    }
 }
